+++ Demo/Demo/DemoEditorScreen.swift
@@ -54,11 +54,6 @@ struct DemoEditorScreen: View {
                 }
             }
         }
-        .overlay(content: {
-            Button("DELETE") {
-                context.trigger(.deleteText(in: .init(location: 2, length: 4)))
-            }
-        })
         .frame(minWidth: 500)
         .focusedValue(\.richTextContext, context)
         .toolbarRole(.automatic)