+++ RELEASE_NOTES.md
@@ -18,7 +18,9 @@ Reach out if you think that this version changes too much for a minor bump, or i
 
 * `RichTextAction` has a new `deleteSelectedText` action.
 * `RichTextAction` has a new `deleteText(at:)` action.
-* `RichTextView` has a new `deleteAction(at:)` function.
+* `RichTextAction` has a new `replaceSelectedText(with:)` action.
+* `RichTextAction` has a new `replaceText(at:with:)` action.
+* `RichTextView` has a new `deleteText(at:)` function.
 * `RichTextView.Configuration` has a new `isScrollBarsVisible` property for macOS.
 
 ### 💡 Adjustments